Debugging Crucials in UI

+++ lib/timeline/services/expense_data.dart
@@ -1,6 +1,6 @@
 class ExpenseData{
   String tag="Undefined";
-  int amount=0;
+  double amount=0;
   String? remarks="None";
   String time="2022-11-17 13:33:0.245";
   ExpenseData({required this.tag, required this.amount, this.remarks, required this.time});
